requirements.txt
args==0.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clint==0.5.1
console==0.9902
ezenv==0.91
idna==2.10
joblib==0.16.0
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.0
plotly==4.9.0
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
Pygments==2.6.1
PyInquirer==1.0.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
Sastrawi==1.0.1
six==1.15.0
tqdm==4.48.2
tweepy==3.9.0
urllib3==1.25.10
wcwidth==0.2.5
